Create docker-compose.yaml
The root user is configured to have full administrative access to the MongoDB server. When MONGODB_ROOT_PASSWORD is not specified the server allows unauthenticated and unrestricted access.

You can create a user with restricted access to a database while starting the container for the first time. To do this, provide the 
MONGODB_USERNAME, MONGODB_PASSWORD and MONGODB_DATABASE environment variables.